* Converted download function to use open-uri instead of net/http.  * Got the cross-compilation part of the win32 build working. Still need to    finish up the actual gem-building.
--HG--
extra : convert_revision : svn%3A9e37cee1-ff4c-dd11-94a8-00304827d434/trunk%4088

+++ win32.rb
@@ -13,6 +13,7 @@
 # * Aaron Patterson
 #   http://tenderlovemaking.com/2008/11/21/cross-compiling-ruby-gems-for-win32/
 
+require 'rake'
 require 'pathname'
 require 'rbconfig'
 
@@ -29,9 +30,9 @@ XCOMPILER_SRC  = XCOMPILER_DIR + "ruby-#{RUBYVERSION}"
 XCOMPILER_BIN  = XCOMPILER_DIR + 'bin'
 XCOMPILER_RUBY = XCOMPILER_BIN + 'ruby.exe'
 
-NEW_ALT_SEPARATOR = '"\\\\\" ' + ?\
+TAR_OPTS = { :compression => :gzip }
 
-CONFIGURE_CMD = %w[
+CONFIGURE_CMD = %W[
 	env
 		ac_cv_func_getpgrp_void=no
 		ac_cv_func_setpgrp_void=yes
@@ -41,41 +42,95 @@ CONFIGURE_CMD = %w[
     ./configure
 	    --host=i386-mingw32
 	    --target=i386-mingw32
-	    --build=#{RUBY_PLATFORM}
+	    --build=#{Config::CONFIG['host']}
 	    --prefix=#{XCOMPILER_DIR}
 ]
+
+### Archive::Tar::Reader#extract (as of 0.9.0) is broken w.r.t.
+### permissions, so we have to do this ourselves.
+def untar( tarfile, targetdir )
+	targetdir = Pathname( targetdir )
+	raise "No such directory: #{targetdir}" unless targetdir.directory?
+
+	reader = Archive::Tar::Reader.new( tarfile.to_s, TAR_OPTS )
+	
+	mkdir_p( targetdir )
+	reader.each( true ) do |header, body|
+		path = targetdir + header[:path]
+		# trace "Header is: %p" % [ header ]
+
+		case header[:type]
+		when :file
+			trace "  #{path}"
+			path.open( File::WRONLY|File::EXCL|File::CREAT|File::TRUNC, header[:mode] ) do |fio|
+				bytesize = header[:size]
+				fio.write( body[0,bytesize] )
+			end
+
+		when :directory
+			trace "  #{path}"
+			path.mkpath
+			
+		when :link
+			linktarget = targetdir + header[:dest]
+			trace "  #{path} => #{linktarget}"
+			path.make_link( linktarget.to_s )
+
+		when :symlink
+			linktarget = targetdir + header[:dest]
+			trace "  #{path} -> #{linktarget}" 
+			path.make_symlink( linktarget )
+		end
+	end
+	
+end
+
 
 begin
 	require 'archive/tar'
 	
 	namespace :win32 do
-		directory XCOMPILER_DIR
+		directory XCOMPILER_DIR.to_s
 		
 		file XCOMPILER_DL => XCOMPILER_DIR do
-			download RUBY_DL_URI, XCOMPILER_DL
+			# openuri can't handle this -- passive ftp required?
+			run 'wget', '-O', XCOMPILER_DL, RUBY_DL_URI
 		end
 
-		directory XCOMPILER_SRC
+		directory XCOMPILER_SRC.to_s
 		task XCOMPILER_SRC => [ XCOMPILER_DIR, XCOMPILER_DL ] do
-			Archive::Tar.extract( XCOMPILER_DL, XCOMPILER_DIR, :compression => :gzip ) or
-				fail "Extraction of %s failed." % [ XCOMPILER_DL ]
+			if XCOMPILER_SRC.exist?
+				trace "Rake fails. #{XCOMPILER_SRC} already exists."
+			else
+				untar( XCOMPILER_DL, XCOMPILER_DIR )
+			end
 		end
 
 		file XCOMPILER_RUBY => XCOMPILER_SRC do
 			Dir.chdir( XCOMPILER_SRC ) do
-				File.open( 'Makefile.in.new', IO::CREAT|IO::WRONLY|IO::EXCL ) do |ofh|
-					File.each_line( 'Makefile.in' ) do |line|
-						line.sub!( /ALT_SEPARATOR = "\\\\"/, NEW_ALT_SEPARATOR )
-						ofh.write( line )
+				unless File.exist?( 'Makefile.in.orig' )
+					File.open( 'Makefile.in.new', IO::CREAT|IO::WRONLY|IO::EXCL ) do |ofh|
+						IO.readlines( 'Makefile.in' ).each do |line|
+							next if line.include?( 0.chr )
+							trace "  copying line: %p" % [ line ]
+							line.sub!( /ALT_SEPARATOR = ".*?"/, "ALT_SEPARATOR = 92.chr" )
+							ofh.write( line )
+						end
 					end
-				end
 
-				mv 'Makefile.in.new', 'Makefile.in'
+					mv 'Makefile.in', 'Makefile.in.orig'
+					mv 'Makefile.in.new', 'Makefile.in'
+				end
 				
 				run *CONFIGURE_CMD
+				run 'make', 'ruby'
+				run 'make', 'rubyw.exe'
+				run 'make', '-n', 'install'
 			end
 		end
 
+		desc "Build a binary Gem for Win32 systems, installing a cross " +
+		     "compiled Ruby if necessary"
 		task :build => XCOMPILER_RUBY do
 			log "Building..."
 		end
@@ -87,8 +142,11 @@ rescue LoadError => err
 	end
 	
 	namespace :win32 do
+		desc "Build a binary Gem for Win32 systems, installing a cross " +
+		     "compiled Ruby if necessary"
 		task :build => :no_win32_build
 	end
 			
 end
+
